+++ src/main/java/com/kosa/realestate/realestates/controller/RealEstateSaleController.java
@@ -1,18 +1,19 @@
 package com.kosa.realestate.realestates.controller;
 
 import java.util.List;
-
+import java.util.Map;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import com.kosa.realestate.realestates.model.RealEstateSaleDTO;
 import com.kosa.realestate.realestates.model.RealEstateWithSaleDTO;
 import com.kosa.realestate.realestates.service.IRealEstateSaleService;
-
 import lombok.RequiredArgsConstructor;
 
 @Controller
@@ -56,7 +57,23 @@ public class RealEstateSaleController {
       model.addAttribute("realEstatePrice", realEstatePrice);
       return "detail";
   }
-  
-  
+  //검색 조건에 따른 매물 검색
+  @PostMapping("/search/byCriteria")
+  @ResponseBody
+  public List<RealEstateWithSaleDTO> selectRealEstateWithSalesByCondition(
+      @RequestBody(required = false) Map<String, Object> searchCriteria) {
+    // 지역구 조건
+    String districtName = searchCriteria != null && searchCriteria.get("districtName") != null ? (String) searchCriteria.get("districtName") : null;
+    // 동 조건의 여부
+    String neighborhoodName = searchCriteria != null && searchCriteria.get("neighborhoodName") != null ? (String) searchCriteria.get("neighborhoodName") : null;
+    // 가격 조건
+    Integer minPrice = searchCriteria != null && searchCriteria.get("PriceMin") != null ? (Integer) searchCriteria.get("PriceMin") : null;
+    Integer maxPrice = searchCriteria != null && searchCriteria.get("PriceMax") != null ? (Integer) searchCriteria.get("PriceMax") : null;
+    //전용면적 조건
+    Integer minExclusiveSize = searchCriteria != null && searchCriteria.get("ExclusiveSizeMin") != null ? (Integer) searchCriteria.get("ExclusiveSizeMin") : null;
+    Integer maxExclusiveSize = searchCriteria != null && searchCriteria.get("ExclusiveSizeMax") != null ? (Integer) searchCriteria.get("ExclusiveSizeMax") : null;
+    
+    return realEstateSaleService.selectRealEstateWithSalesByCondition(districtName, neighborhoodName, minPrice, maxPrice, minExclusiveSize, maxExclusiveSize);
+  }
   
 }
