+++ src/main/java/com/kosa/realestate/realestates/repository/RealEstateSaleRepository.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Map;
 
 import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
 import com.kosa.realestate.realestates.model.RealEstateSaleDTO;
 import com.kosa.realestate.realestates.model.RealEstateWithSaleDTO;
 
@@ -19,7 +20,7 @@ public interface RealEstateSaleRepository {
   //동 리스트 조회
   List<Map<String, Object>> getAllNeighborhood(int destrictId);
   //검색조건으로 매매기록 조회하기
-  List<RealEstateWithSaleDTO> selectRealEstateWithSalesByCondition();
+  List<RealEstateWithSaleDTO> selectRealEstateWithSalesByCondition(@Param("districtName") String districtName, @Param("neighborhoodName")String neighborhoodName,@Param("salePrice") Map<String, Object> salePrice,@Param("exclusiveArea") Map<String, Object> exclusiveArea);
   
   List<RealEstateWithSaleDTO> getRealEstateDetail(int salesId);
   List<RealEstateSaleDTO> getRealEstatePrice(int salesId);
