+++ src/main/java/com/kosa/realestate/realestates/service/IRealEstateSaleService.java
@@ -12,7 +12,9 @@ public interface IRealEstateSaleService {
   
   List<Map<String, Object>> getAllDestrictId();
   List<Map<String, Object>> getAllNeighborhood(int destrictId);
-  
+  // 검색조건으로 매매기록 조회하기
+  List<RealEstateWithSaleDTO> selectRealEstateWithSalesByCondition(String districtName, String neighborhoodName, int minPrice, int maxPrice, int minExclusiveSize, int maxExclusiveSize);
+
   List<RealEstateWithSaleDTO> getRealEstateDetail(int salesId);
   List<RealEstateSaleDTO> getRealEstatePrice(int salesId);
 }
