+++ src/main/resources/templates/map.html
@@ -158,6 +158,7 @@
             }
         });
 
+				
 	</script>
 </body>
